refactor: 검색 페이지 파일구조 리팩토링
